+++ test/cli.js
@@ -1,9 +1,8 @@
-'use strict'
+import path from 'node:path'
 
-const access = require('./helpers/access')
-const { cleanupOutputDir, tempOutputDir } = require('./helpers/describe_installer')
-const path = require('path')
-const spawn = require('../src/spawn')
+import access from './helpers/access.js'
+import { cleanupOutputDir, tempOutputDir } from './helpers/describe_installer.js'
+import spawn from '../src/spawn.js'
 
 function runCLI (options) {
   const args = [