+++ test/helpers/describe_installer.js
@@ -1,48 +1,46 @@
-'use strict'
+import { expect } from 'chai'
+import fs from 'node:fs/promises'
+import os from 'node:os'
+import path from 'node:path'
+import tmp from 'tmp-promise'
 
-const { expect } = require('chai')
-const fs = require('node:fs/promises')
-const os = require('node:os')
-const path = require('node:path')
-const tmp = require('tmp-promise')
+import installer from '../../src/installer.js'
 
-const installer = require('../..')
-
-module.exports = function describeInstaller (description, installerOptions, itDescription, itFunc) {
+export default function describeInstaller (description, installerOptions, itDescription, itFunc) {
   describe(description, () => {
-    const outputDir = module.exports.tempOutputDir(installerOptions.dest)
-    const options = module.exports.testInstallerOptions(outputDir, installerOptions)
+    const outputDir = tempOutputDir(installerOptions.dest)
+    const options = testInstallerOptions(outputDir, installerOptions)
 
     before(() => installer(options))
 
     it(itDescription, () => itFunc(outputDir))
 
-    module.exports.cleanupOutputDir(outputDir)
+    cleanupOutputDir(outputDir)
   })
 }
 
-module.exports.describeInstallerWithException = function describeInstallerWithException (description, installerOptions, errorRegex) {
+export function describeInstallerWithException (description, installerOptions, errorRegex) {
   describe(description, () => {
-    const outputDir = module.exports.tempOutputDir()
-    module.exports.cleanupOutputDir(outputDir)
+    const outputDir = tempOutputDir()
+    cleanupOutputDir(outputDir)
 
     it('throws an error', () => {
-      const options = module.exports.testInstallerOptions(outputDir, installerOptions)
+      const options = testInstallerOptions(outputDir, installerOptions)
       return installer(options)
         .catch(error => expect(error.message).to.match(errorRegex))
     })
   })
 }
 
-module.exports.cleanupOutputDir = function cleanupOutputDir (outputDir) {
+export function cleanupOutputDir (outputDir) {
   after(() => fs.rm(outputDir, { recursive: true, force: true }))
 }
 
-module.exports.tempOutputDir = function tempOutputDir (customDir) {
+export function tempOutputDir (customDir) {
   return customDir ? path.join(os.tmpdir(), customDir) : tmp.tmpNameSync({ prefix: 'electron-installer-debian-' })
 }
 
-module.exports.testInstallerOptions = function testInstallerOptions (outputDir, installerOptions) {
+export function testInstallerOptions (outputDir, installerOptions) {
   return {
     rename: debFile => {
       return path.join(debFile, '<%= name %>_<%= arch %>.deb')